Scucu/explicit subscription in deployment script (#38)
* Provide explicit subscription to install script

* Allow user to provide non-default subscription during deployment

The subscription can be provided as an optional argument. Otherwise, it is inferred from teh current account. If multiple subscriptions are found, the user can choose which one to use for the deployment.
Also, no longer provide tenant explicitly to avoid discrepancies with the subscription id. Instead, derive tenant from current subscription.
Remove a few warnings

* Update documentation about how to choose subscription for deployment

* Clarifications on how monitoring ADF pipelines post-deployment

* Add disclaimer for building on Windows systems

* Enhancements to deployment script

* Remove default value for app service since it must be globally unique